+++ src/main/java/com/highstreet/checkout/services/CheckoutService.java
@@ -1,12 +1,14 @@
 package com.highstreet.checkout.services;
 
+import java.math.BigDecimal;
+
 import com.highstreet.checkout.model.Item;
 
 
 
 public interface CheckoutService {
 
-    Double total();
+    BigDecimal total();
 
     void scan(Item item);
